requirements.txt
autopep8==1.6.0
cycler==0.11.0
imageio==2.10.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.4
Pillow==8.4.0
plotly==5.3.1
pycodestyle==2.8.0
pyparsing==3.0.5
python-dateutil==2.8.2
PyWavelets==1.2.0
richdem==0.3.4
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
simpleai==0.8.3
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2